pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyGithub>=2.1.1",
    "fastmcp>=0.1.0",
    "pypinyin>=0.50.0",
]
